+++ cs/Markdown/Lexer/Lexer.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using Markdown.Interfaces;
+using Markdown.Tokens;
 
-namespace Markdown
+namespace Markdown.Lexer
 {
     public class Lexer : ILexer
     {
