+++ cs/Markdown/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Markdown.TokenRenderer;
 
 namespace Markdown
 {
@@ -6,8 +7,8 @@ namespace Markdown
     {
         private static void Main(string[] args)
         {
-            var lexer = new Lexer();
-            var tokenParser = new TokenParser();
+            var lexer = new Lexer.Lexer();
+            var tokenParser = new TokenParser.TokenParser();
             var htmlTokenRenderer = new HtmlTokenRenderer();
             var md = new Md(lexer, tokenParser, htmlTokenRenderer);
             Console.WriteLine(md.Render("__Выделенный двумя символами текст__"));
