+++ cs/Markdown/Tokens/Token.cs
@@ -1,4 +1,4 @@
-﻿namespace Markdown
+﻿namespace Markdown.Tokens
 {
     public readonly struct Token
     {
