+++ cs/Markdown/Tokens/TokenType.cs
@@ -1,4 +1,4 @@
-﻿namespace Markdown
+﻿namespace Markdown.Tokens
 {
     public enum TokenType
     {
